\name{expSBM-package}
\alias{expSBM-package}
\alias{expSBM}
\docType{package}
\title{
\packageTitle{expSBM}
}
\description{
\packageDescription{expSBM}
}
\author{
Riccardo Rastelli <riccardoras@gmail.com>

Michael Fop <michael.fop@ucd.ie>
}
\references{
R. Rastelli and M. Fop (2019) "A dynamic stochastic blockmodel for interaction lengths", \url{https://arxiv.org/abs/1901.09828}
}
\keyword{ package }
